+++ app/src/main/java/com/battlelancer/seriesguide/movies/MoviesNowFragment.kt
@@ -283,22 +283,22 @@ class MoviesNowFragment : Fragment() {
             }
         }
 
-    private val traktFriendsHistoryCallbacks: LoaderManager.LoaderCallbacks<List<NowItem>> =
-        object : LoaderManager.LoaderCallbacks<List<NowItem>> {
-            override fun onCreateLoader(id: Int, args: Bundle?): Loader<List<NowItem>> {
-                return TraktFriendsMovieHistoryLoader(activity)
+    private val traktFriendsHistoryCallbacks: LoaderManager.LoaderCallbacks<List<NowItem>?> =
+        object : LoaderManager.LoaderCallbacks<List<NowItem>?> {
+            override fun onCreateLoader(id: Int, args: Bundle?): Loader<List<NowItem>?> {
+                return TraktFriendsMovieHistoryLoader(requireContext())
             }
 
             override fun onLoadFinished(
-                loader: Loader<List<NowItem>>,
-                data: List<NowItem>
+                loader: Loader<List<NowItem>?>,
+                data: List<NowItem>?
             ) {
                 adapter.setFriendsRecentlyWatched(data)
                 isLoadingFriends = false
                 showProgressBar(false)
             }
 
-            override fun onLoaderReset(loader: Loader<List<NowItem>>) {
+            override fun onLoaderReset(loader: Loader<List<NowItem>?>) {
                 // clear existing data
                 adapter.setFriendsRecentlyWatched(null)
             }